Tips.ApiMessage - Added Response helper methods to add a notification, has errors, and set status.  I wanted to remove the direct dependency on the HttpStatusCode.  I'll need to revisit the status methods.
Replaced the mapping logic with a GenericMapper that uses the json serializer to serialize one type and deserialize to another type.

Updated RulesEngine.ProcessRules parameters to use the request and the response.  Added a couple more rules to illustrate the concept more.

Pulled the notification factory methods in the handlers back into the handle method for reevaluation to see if there is a better pattern.