Tips.ApiMessage - Added Guard.AgainstNull to standardize enforcing parameters are not null.  While I try to use nameof when it makes sense, that method only grabs the last property name in the chain when I wanted the full chain.  While I could use string interpolation to build the string name, it makes it more difficult when you are searching for the error message based on logs.
Replaced the GenericMapper with the TodoItemMapper.  Turns out I accidently removed the mapping for the Update methods and I can't use JsonSerializer to populate the models.  That feature is available in the Newtonsoft, but not .NET Core json library.

The rules engine was the most effected.  Adding the Guard.AgainstNull.  Added mapping per field to simulate the actual mapping in a rules engine.