Tips.ApiMessage - Updated rules to use Request<TodoItem> and Response<TodoItemEntity>.  This allows the rules to change context only fields like the Reviewer in the new rule added TodoItemReviewerRule.
Updated the BaseRule to mark rules as Skipped to reinforce the choice to skip the rule vs rules that have yet to be processed.

I especially liked that I could put a breakpoint in the RulesEngine at the end of the loop and see every rule that passed, failed, skipped.

Updated the repositories to no longer return a response since there was no mapping necessary.