Tips.ApiMessage - Updated all NuGet packages.  Should have done it separately, but I ran into some errors and it helped knowing it wasn't the problem.
Separated logic into projects based on dependencies.  I want the TodoItems to be in it's own dedicated assembly.  The endpoint project (Tips.ApiMessage) should take advantage of the Pipeline project to route to the TodoItems handlers.  The Piperline and Rules projects consist of the basic framework required to support both concepts.

No logic changes were made.  Just a lot of file moving and namespace updates.

However, I did change the BaseRule to handle the guard against null logic inline to avoid an unnecessary dependency on the Guard class.

Not sure why, but I had to update the log4net.config to reference the log file 3 more folders up.  Maybe something with the log4net nuget update?