Tips.ApiMessage - Added the NotFoundNotification class as a special notification type used to denote that the item was not found when getting, updating, or deleting a single item.  It's not directly tied/coupled to the API concept, but it supports it.
Added a IsNotFound method that returns true if any of the notifications were an instance of that type.  Additional types can be created to support other concepts.  Additional checks can be added to the Response directly or via extension methods.

Removed the Status methods by refactoring the setting of the HttpStatusCodes in the controller only when needed.