Tips.ApiMessage - Created a ProblemDetailFactory and refactored the bad request and internal server error versions into it.  It is a dependency within the Middleware class, so it too must be injected as a singleton.
Renamed the configuration file and reference in the appsettings.json to ProblemDetailConfiguration given it's now used in the ProblemDetailFactory.

Response no longer needs the Status or TraceId.  Status and TraceId are handled with the ProblemDetails as needed.  TraceId is also handled in the logs so it's everywhere it needs to be.