Tips.ApiMessage - Added log4net.xsd and added schema reference to remove the 'log4net' element is not declared.
Updated exceptions to use ArgumentNullException and InvalidOperationException as recommended in Exceptions Best Practices.
https://docs.microsoft.com/en-us/dotnet/standard/exceptions/best-practices-for-exceptions

Refactored the Guard class to it's own GuardClauses project.  I only use it in the TodoItems project.  While it could be updated to guard against InvalidOperationExceptions, I only have one case for that in hte ApiKeyHandlerMiddleware.  While I could use the Guard.AgainstNull in the Rules project as well, I keep the dependencies to a minimum for easy reuse.