Tips.ApiMessage - Updated Notification to support child Notifications.  Added an nested notification example to the contrived Reviewer rule.
Updated Response to no longer support Add and AddRange.  They were originally added to reduce nested coupling following the principle law of demeter.  However, the only additional coupling here is to the list class which is part of the .NET framework which is a reasonable coupling to have.  Updated references to .Add with .Notifications.Add and .AddRange with .Notifications.AddRange.