Tips.ApiMessage - Decided to implement and IBaseRule interface for BaseRule.  Exposed Status and RuleStatusType.  Added interface default implementations for basic methods like IsPassed.  Renamed methods like RuleSkipped to Skip to enforce that the method was taking action to skip, etc.
While this does make it easier to Mock (Moq) for unit tests, I just as easily could have added "virtual" to any method I needed to setup for unit tests.  I think extracting the interface makes the desired contract clearer and the implementation cleaner.  I did like hiding the RuleStatusType as a private class, though this is a minor gripe for the cleaner interface.

Updated the RulesEngine to only maintain/process the single rules list vs. 2 lists when a simple LINQ statement provided the same results with more info by having access to all rules.  Less chance for something to be missed.