Tips.ApiMessage - Added ServiceCollectionExtensionsTest.  In order to test this fully would require changing classes from internal to public which defeats the purpose.
Alternatively, we could convert the ServiceCollectionExtension to a concrete class, register in the Program class, so it could be injected in the Startup class.

In general, this is covered via integration tests.  Create a unit test if this is critical to your efforts, defects appear in this code, or any other reason warranting unit tests over an integration test.

---

Renamed appsettings.json to appsettings.test.json in each test project to make it clear it's not the actual appsettings.json.

---

Refactored AssertType methods into the centralized method DependencyRegistrarSupport.AssertServiceIsInstanceOfType

Refactored the AddScope of FakeLogger, and FakeLogger implementation into the Support project.  Added DependencyRegistrarSupport.AddScopedLogger.