Tips.ApiMessage - Renamed the support method Populate to PopulateTodoItems to make the intent clearer.
Added a support helper method to GetTodoItem.  While I could call the other GetTodoItemRequestHandler to return the item, this unit test is meant to only test the DeleteTodoItemRequestHandler.  Thankfully, this was easily resolved with this helper.

Updated DeleteTodoItemRequestHandlerTest to verify the TodoItem was in the database before the action, and not there after the delete.