Added swagger ui / swashbuckle.
Added missing xml comments to the warnings to ignore.  Added GenerateDocumentationFile true, so we don't need to specify a path.

Added xml comments on the public endpoints.  Note that the [FromServices] parameters are skipped.

Added a Swagger specific api-key.  This would allow us to restrict or flag these entries to do whatever we want with the records such as return dummy responses, so authenticated users could still try out the api endpoints.

I'm using the hard-coded "x-api-key" for now.  I'll see what I can do about that later.

I added an example on how to remove the swagger Try it out feature.  I implemented the bare minimum required to get authenticate the api-key for each method.

I wanted the API endpoints exposed to give an idea of how they are used, but restrict the Try it out now feature to those authorized with the sample api-key or use their own if authorized.