Added a CreateTodoItem specific rule.
Changed the declaration order of the rules factory injection to be consistent with how they are called in code.  This is just a style preference.

Updated CreateTodoItemRequestHandlerTest to test the new CreateRulesFactory is called when appropriate.

Replaced the UpdateTodoItemRequestHandlerTest to actually call the UpdateTodoItemRequestHandler.  It was calling the Create.  Most likely I copied and pasted, never got back to it and checked in missing this in error or made a note to update which got lost.  Regardless, it's fixed now.  :)